Python Modulo User Interface
This code is a collection of different calculations and outputs performed based on user input. It has multiple sections that handle various calculations using the modulo operator and other mathematical operations. 
The code starts by displaying a welcome message to the user.
The First Code
It prompts the user to enter a number of seconds to calculate the corresponding minutes and seconds. If the user input is less than 60, it raises a value error. The code then calculates the number of minutes and seconds using floor division (//) and modulo (%). It prints the calculated minutes and seconds.
The Second Code
 Ask the user to enter two hours between 1 and 12. If the input values are not within the specified range, it raises a value error. The code calculates the future hour by adding the two hours and performing modulo 12. It prints the future hour.
The Third Code
 Find the last digit of 2 raised to the power of a number entered by the user.
If the user input is negative e.g., -1 or -2, it raises a value error. The code calculates the last digit using modulo 10. then prints the last digit.
The Fourth Code
 finds the last two digits of 2 raised to the power of a user-inputted number. If the user input is negative, it raises a value error. The code calculates the last two digits using modulo 100. Then prints the last two digits.
The Fifth Code
  Asks the user to enter a digit and a power to find the last digit of the number when raised to the power of 2. If the power or digit is negative, it raises a value error. The code calculates the result and prints it. Finally, the code handles any value errors that may occur during user input and prints the respective error message.

In conclusion, the code interacts with the user to perform various calculations using mathematical operations and the modulo operator, it displays the results based on the provided input by user.

Thanks for viewing.